requirements.txt
python-dotenv
streamlit
sqlalchemy
bcrypt
pyjwt
requests
pytest
reportlab
matplotlib
pandas
PyPDF2
